Handling more pull request's comments
