+++ frontend/src/store/ApplicationState.ts
@@ -1,5 +1,5 @@
-import type { FamApplication } from 'fam-api';
 import { computed, ref, reactive } from 'vue';
+import type { FamApplication } from 'fam-api';
 
 // The applications the user has access to administer
 export const applicationsUserAdministers = ref<FamApplication[]>([]);
@@ -37,12 +37,3 @@ export const selectedApplicationDisplayText = computed(() => {
         return '';
     }
 });
-
-export const useNotificationMessage = reactive({
-    notificationMsg: '',
-    isNotificationVisible: false,
-})
-
-export const useErrorDialog = reactive({
-    isErrorVisible: false,
-})
