+++ frontend/src/store/NotificationState.ts
@@ -1,0 +1,10 @@
+import { reactive } from "vue"
+
+export const useNotificationMessage = reactive({
+    notificationMsg: '',
+    isNotificationVisible: false,
+})
+
+export const useErrorDialog = reactive({
+    isErrorVisible: false,
+})
